+++ slashtags/testing/button_menus.py
@@ -8,7 +8,7 @@ from redbot.vendored.discord.ext import menus
 
 from ..http import Button, ButtonStyle, Component, InteractionButton
 
-__all__ = ("PageSource", "ButtonMenu", "BaseButtonMenu", "menu")
+__all__ = ("PageSource", "ButtonMenu", "BaseButtonMenu", "menu", "ButtonMenuMixin")
 
 log = logging.getLogger("red.phenom4n4n.slashtags.testing.button_menus")
 
@@ -47,11 +47,11 @@ class MenuButton(menus.Button):
         self.style = style
 
 
-class BaseButtonMenu(menus.MenuPages, inherit_buttons=False):
-    def __init__(self, source: menus.PageSource, *, custom_id: str = None, **kwargs):
+class ButtonMenuMixin:
+    def __init__(self, *args, **kwargs):
         # kwargs["clear_reactions_after"] = False
-        super().__init__(source, **kwargs)
-        self.custom_id = custom_id
+        self.custom_id = kwargs.pop("custom_id", None)
+        super().__init__(*args, **kwargs)
         self._buttons_closed = True
         self._components = []
 
@@ -59,58 +59,13 @@ class BaseButtonMenu(menus.MenuPages, inherit_buttons=False):
     def __tasks(self):
         return self._Menu__tasks
 
-    async def show_checked_page(self, page_number: int, button: InteractionButton):
-        max_pages = self._source.get_max_pages()
-        try:
-            if max_pages is None or page_number < max_pages and page_number >= 0:
-                # If it doesn't give maximum pages, it cannot be checked
-                await self.show_page(page_number, button)
-            elif page_number >= max_pages:
-                await self.show_page(0, button)
-            else:
-                await self.show_page(max_pages - 1, button)
-        except IndexError:
-            # An error happened that can be handled, so ignore it.
-            pass
-
-    async def show_page(
-        self, page_number: int, button: InteractionButton, *, recalculate_components: bool = True
-    ):
-        page = await self._source.get_page(page_number)
-        self.current_page = page_number
-        kwargs = await self._get_kwargs_from_page(page)
-        if recalculate_components:
-            kwargs["components"] = self._get_components()
-        await button.update(**kwargs)
-
-    async def send_initial_message(
-        self,
-        ctx: commands.Context,
-        channel: discord.TextChannel,
-        *,
-        reply: bool = False,
-        mention_author: bool = False,
-    ):
-        page = await self._source.get_page(0)
-        kwargs = await self._get_kwargs_from_page(page)
-        if not self.custom_id:
-            self.custom_id = str(ctx.message.id)
-        if reply:
-            kwargs["reference"] = ctx.message.to_reference(fail_if_not_exists=False)
-            kwargs["mention_author"] = mention_author
-        return await self._send(ctx, **kwargs)
+    def _get_component_from_emoji(self, emoji: discord.PartialEmoji) -> Button:
+        return Button(style=ButtonStyle.grey, custom_id=f"{self.custom_id}-{emoji}", emoji=emoji)
 
     def _get_components(self) -> List[Component]:
         components = []
         for emojis in chunks(list(self.buttons.keys()), 5):
-            buttons = [
-                Button(
-                    style=ButtonStyle.grey,
-                    custom_id=f"{self.custom_id}-{emoji}",
-                    emoji=emoji,
-                )
-                for emoji in emojis
-            ]
+            buttons = [self._get_component_from_emoji(emoji) for emoji in emojis]
             components.append(Component(components=buttons))
         return components
 
@@ -150,67 +105,6 @@ class BaseButtonMenu(menus.MenuPages, inherit_buttons=False):
             asyncio.create_task(button.send("You cannot use this menu.", hidden=True))
             return False
         return button.custom_id.startswith(self.custom_id)
-
-    async def _edit_button_components(
-        self, button: InteractionButton, components: List[Component]
-    ):
-        page = await self._source.get_page(self.current_page)
-        kwargs = await self._get_kwargs_from_page(page)
-        kwargs["components"] = components
-        await button.update(**kwargs)
-
-    async def _edit_message_components(self, components: List[Component]):
-        route = discord.http.Route(
-            "PATCH",
-            "/channels/{channel_id}/messages/{message_id}",
-            channel_id=self.message.channel.id,
-            message_id=self.message.id,
-        )
-        data = {"components": [c.to_dict() for c in components]}
-        await self.bot._connection.http.request(route, json=data)
-
-    async def close_buttons(self, button: InteractionButton = None):
-        if self._buttons_closed:
-            return
-        if button:
-            page = await self._source.get_page(self.current_page)
-            kwargs = await self._get_kwargs_from_page(page)
-            await button.update(**kwargs, components=[])
-        else:
-            await self._edit_message_components([])
-        self._buttons_closed = True
-
-    async def change_source(self, source: menus.PageSource, button: InteractionButton):
-        if not isinstance(source, menus.PageSource):
-            raise TypeError("Expected {0!r} not {1.__class__!r}.".format(PageSource, source))
-
-        self._source = source
-        self.current_page = 0
-        if self.message is not None:
-            await source._prepare_once()
-            await self.show_page(0, button)
-
-    def add_button(self, button, *, react=False, interaction: InteractionButton = None):
-        self._buttons[button.emoji] = button
-
-        if react:
-            if self.__tasks:
-
-                async def wrapped():
-                    # Add the component
-                    self.buttons[button.emoji] = button
-                    components = self._get_components()
-                    if interaction:
-                        await self._edit_button_components(interaction, components)
-                    else:
-                        await self._edit_message_components(components)
-
-                return wrapped()
-
-            async def dummy():
-                raise menus.MenuError("Menu has not been started yet")
-
-            return dummy()
 
     async def start(self, ctx, *, channel=None, wait=False):
         # Clear the buttons cache and re-compute if possible.
@@ -337,6 +231,110 @@ class BaseButtonMenu(menus.MenuPages, inherit_buttons=False):
             log.exception(
                 f"An error occured while updating {type(self).__name__} menu.", exc_info=error
             )
+
+
+class BaseButtonMenu(ButtonMenuMixin, menus.MenuPages, inherit_buttons=False):
+    async def show_checked_page(self, page_number: int, button: InteractionButton):
+        max_pages = self._source.get_max_pages()
+        try:
+            if max_pages is None or page_number < max_pages and page_number >= 0:
+                # If it doesn't give maximum pages, it cannot be checked
+                await self.show_page(page_number, button)
+            elif page_number >= max_pages:
+                await self.show_page(0, button)
+            else:
+                await self.show_page(max_pages - 1, button)
+        except IndexError:
+            # An error happened that can be handled, so ignore it.
+            pass
+
+    async def show_page(
+        self, page_number: int, button: InteractionButton, *, recalculate_components: bool = True
+    ):
+        page = await self._source.get_page(page_number)
+        self.current_page = page_number
+        kwargs = await self._get_kwargs_from_page(page)
+        if recalculate_components:
+            kwargs["components"] = self._get_components()
+        await button.update(**kwargs)
+
+    async def send_initial_message(
+        self,
+        ctx: commands.Context,
+        channel: discord.TextChannel,
+        *,
+        reply: bool = False,
+        mention_author: bool = False,
+    ):
+        page = await self._source.get_page(0)
+        kwargs = await self._get_kwargs_from_page(page)
+        if not self.custom_id:
+            self.custom_id = str(ctx.message.id)
+        if reply:
+            kwargs["reference"] = ctx.message.to_reference(fail_if_not_exists=False)
+            kwargs["mention_author"] = mention_author
+        return await self._send(ctx, **kwargs)
+
+    async def _edit_button_components(
+        self, button: InteractionButton, components: List[Component]
+    ):
+        page = await self._source.get_page(self.current_page)
+        kwargs = await self._get_kwargs_from_page(page)
+        kwargs["components"] = components
+        await button.update(**kwargs)
+
+    async def _edit_message_components(self, components: List[Component]):
+        route = discord.http.Route(
+            "PATCH",
+            "/channels/{channel_id}/messages/{message_id}",
+            channel_id=self.message.channel.id,
+            message_id=self.message.id,
+        )
+        data = {"components": [c.to_dict() for c in components]}
+        await self.bot._connection.http.request(route, json=data)
+
+    async def close_buttons(self, button: InteractionButton = None):
+        if self._buttons_closed:
+            return
+        if button:
+            page = await self._source.get_page(self.current_page)
+            kwargs = await self._get_kwargs_from_page(page)
+            await button.update(**kwargs, components=[])
+        else:
+            await self._edit_message_components([])
+        self._buttons_closed = True
+
+    async def change_source(self, source: menus.PageSource, button: InteractionButton):
+        if not isinstance(source, menus.PageSource):
+            raise TypeError("Expected {0!r} not {1.__class__!r}.".format(PageSource, source))
+
+        self._source = source
+        self.current_page = 0
+        if self.message is not None:
+            await source._prepare_once()
+            await self.show_page(0, button)
+
+    def add_button(self, button, *, react=False, interaction: InteractionButton = None):
+        self._buttons[button.emoji] = button
+
+        if react:
+            if self.__tasks:
+
+                async def wrapped():
+                    # Add the component
+                    self.buttons[button.emoji] = button
+                    components = self._get_components()
+                    if interaction:
+                        await self._edit_button_components(interaction, components)
+                    else:
+                        await self._edit_message_components(components)
+
+                return wrapped()
+
+            async def dummy():
+                raise menus.MenuError("Menu has not been started yet")
+
+            return dummy()
 
 
 class ButtonMenu(BaseButtonMenu, inherit_buttons=False):
